Split encrpt file into key & offset files
