+++ lib/offset.rb
@@ -1,0 +1,7 @@
+# def todays_date
+#   Time.now.strftime("%d%m%y")
+# end
+#
+# def offset
+#   date = todays_date.to_i ** 2).to_s.slice(-4..-1)
+# end
