+++ test/offset_test.rb
@@ -1,18 +1,8 @@
 require 'minitest/autorun'
 require 'minitest/pride'
-require './lib/encrypt'
+require './lib/offset'
 
-class EncryptTest < Minitest::Test
-
-  def test_it_exists
-    encrypt = Encrypt.new
-    assert_instance_of Encrypt, encrypt
-  end
-
-  def test_it_generates_five_digit_key
-    encrypt = Encrypt.new
-    assert_equal 5, encrypt.generate_key.length
-  end
+class OffsetTest < Minitest::Test
 
   def test_it_knows_todays_date
     encrypt = Encrypt.new
